requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers
huggingface_hub
soundfile
numpy
sentencepiece
